Changes:
Updated youtube controller to insert a record into the videos table with all the data
The users table updates the user record with the points, number of videos watched, total hours spent watching, and level.

the stats route/controller will return the required information in the proper format. There was an syntax issue there that was solved

AuthStrategies now also insert the user record if no user exists and checks if the auth_token is valid, it will just return a boolean true or false based on if the user should be authorized